Beginning refactoring argument parser to implement extractor script
