+++ src/odbp/cli.py
@@ -11,7 +11,7 @@ import argparse
 import numpy as np
 from typing import Any, Union, TypeAlias
 from .odb_visualizer import OdbVisualizer
-from .util import confirm
+from .util import CLIOptions, confirm, process_input, print_state
 from odbp import __version__
 
 
@@ -39,108 +39,6 @@ class UserOptions():
         self.image_label: str = image_label
         self.toml_config_file: ConfigFiletype = toml_config_file
         self.run_immediate: bool = run_immediate
-
-
-class CLIOptions():
-    """
-    Struct to store cli optoins without repeating
-    """
-    def __init__(self) -> None:
-        self.quit_options: list[str] = ["exit", "quit", "q"]
-        self.quit_help: str = "Exit ODBPlotter"
-        self.quit_options_formatted: str = ", ".join(self.quit_options)
-
-        self.select_options: list[str] = ["select"]
-        self.select_help: str = "Select an hdf5 file (or generate an hdf5 file from a .odb file)"
-        self.select_options_formatted: str = ", ".join(self.select_options)
-
-        self.seed_options: list[str] = ["seed", "mesh"]
-        self.seed_help: str = "Set the Mesh Seed Size"
-        self.seed_options_formatted: str = ", ".join(self.seed_options)
-
-        self.extrema_options: list[str] = ["extrema", "range"]
-        self.extrema_help: str = "Set the upper and lower x, y, and z bounds for plotting"
-        self.extrema_options_formatted: str = ", ".join(self.extrema_options)
-
-        self.time_options: list[str] = ["time"]
-        self.time_help: str = "Set the upper and lower time bounds"
-        self.time_options_formatted: str = ", ".join(self.time_options)
-
-        self.time_sample_options: list[str] = ["sample"]
-        self.time_sample_help: str = "Set the Time Sample for the hdf5 file"
-        self.time_sample_options_formatted: str = ", ".join(self.time_sample_options)
-
-        self.meltpoint_options: list[str] = ["meltpoint", "melt", "point", "temp", "temperature"]
-        self.meltpoint_help: str = "Set the Melting Point and lower Temperature Bound for the hdf5 file"
-        self.meltpoint_options_formatted: str = ", ".join(self.meltpoint_options)
-
-        self.title_label_options: list[str] = ["title", "label"]
-        self.title_label_help: str = "Set the title and label of the output plots"
-        self.title_label_options_formatted: str = ", ".join(self.title_label_options)
-
-        self.directory_options: list[str] = ["dir", "dirs", "directory", "directories"]
-        self.directory_help: str = "Set the source and output directories"
-        self.directory_options_formatted: str = ", ".join(self.directory_options)
-
-        self.process_options: list[str] = ["process", "run", "load"]
-        self.process_help: str = "Actually load the selected data from the file set in select"
-        self.process_options_formatted: str = ", ".join(self.process_options)
-
-        self.angle_options: list[str] = ["angle", "elev", "elevation", "azim", "azimuth", "roll"]
-        self.angle_help: str = "Update the viewing angle"
-        self.angle_options_formatted: str = ", ".join(self.angle_options)
-
-        self.show_all_options: list[str] = ["show-all", "plot-all"]
-        self.show_all_help: str = "Toggle if each time step will be shown in te matplotlib interactive viewer"
-        self.show_all_options_formatted: str = ", ".join(self.show_all_options)
-
-        self.plot_options: list[str] = ["plot", "show"]
-        self.plot_help: str = "Plot each selected timestep"
-        self.plot_options_formatted: str = ", ".join(self.plot_options)
-        
-        self.state_options: list[str] = ["state", "status", "settings"]
-        self.state_help: str = "Show the current state of the settings of the plotter"
-        self.state_options_formatted: str = ", ".join(self.state_options)
-
-        self.help_options: list[str] = ["help", "use", "usage"]
-        self.help_help: str = "Show this menu"
-        self.help_options_formatted: str = ", ".join(self.help_options)
-
-        self.longest_len: int = max(
-                len(self.quit_options_formatted),
-                len(self.select_options_formatted),
-                len(self.seed_options_formatted),
-                len(self.extrema_options_formatted),
-                len(self.time_options_formatted),
-                len(self.time_sample_options_formatted),
-                len(self.meltpoint_options_formatted),
-                len(self.title_label_options_formatted),
-                len(self.directory_options_formatted),
-                len(self.process_options_formatted),
-                len(self.angle_options_formatted)
-                )
-
-    def print_help(self) -> None:
-        print(
-    f"""ODBPlotter Help:
-{self.help_options_formatted.ljust(self.longest_len)} -- {self.help_help}
-{self.quit_options_formatted.ljust(self.longest_len) } -- {self.quit_help}
-{self.select_options_formatted.ljust(self.longest_len)} -- {self.select_help}
-{self.seed_options_formatted.ljust(self.longest_len)} -- {self.seed_help}
-{self.extrema_options_formatted.ljust(self.longest_len)} -- {self.extrema_help}
-{self.time_options_formatted.ljust(self.longest_len)} -- {self.time_help}
-{self.time_sample_options_formatted.ljust(self.longest_len)} -- {self.time_sample_help}
-{self.meltpoint_options_formatted.ljust(self.longest_len)} -- {self.meltpoint_help}
-{self.title_label_options_formatted.ljust(self.longest_len)} -- {self.title_label_help}
-{self.directory_options_formatted.ljust(self.longest_len)} -- {self.directory_help}
-{self.process_options_formatted.ljust(self.longest_len)} -- {self.process_help}
-{self.angle_options_formatted.ljust(self.longest_len)} -- {self.angle_help}
-{self.show_all_options_formatted.ljust(self.longest_len)} -- {self.show_all_help}
-{self.plot_options_formatted.ljust(self.longest_len)} -- {self.plot_help}
-{self.state_options_formatted.ljust(self.longest_len)} -- {self.state_help}"""
-    )
-
-
 
 
 def cli():
@@ -183,6 +81,9 @@ def cli():
 
             elif user_input in cli_options.meltpoint_options:
                 set_meltpoint(state)
+
+            elif user_input in cli_options.low_temp_options:
+                set_low_temp(state)
 
             elif user_input in cli_options.title_label_options:
                 set_title_and_label(state, user_options)
@@ -670,267 +571,8 @@ def process_extrema(keys: tuple[str, str]) -> tuple[float, float]:
     return tuple(results)
 
 
-def process_input(state: OdbVisualizer) -> UserOptions:
-    """
-    The goal is to have a hierarchy of options. If a user passes in an option via a command-line switch, that option is set.
-    If an option is not set by a switch, then the toml input file is used.
-    If an option is not set by the input file (if any), then prompt for it
-    Possibly also include a default config file, like in $HOME/.config? Not sure
-    """
-
-    parser: argparse.ArgumentParser = argparse.ArgumentParser(description="ODBPlotter")
-    parser.add_argument("config_file", nargs="?")
-
-    parser.add_argument("-s", "--hdf-source-directory")
-    parser.add_argument("-b", "--odb-source-directory")
-    parser.add_argument("-r", "--results-directory")
-
-    parser.add_argument("-o", "--odb")
-    parser.add_argument("-m", "--meltpoint")
-    parser.add_argument("-l", "--low-temp", default=300)
-    parser.add_argument("-S", "--seed-size")
-    parser.add_argument("-t", "--time-sample", default=1)
-
-    parser.add_argument("-H", "--hdf")
-
-    parser.add_argument("-T", "--title")
-    parser.add_argument("-L", "--label")
-
-    parser.add_argument("--low-x")
-    parser.add_argument("--high-x")
-    parser.add_argument("--low-y")
-    parser.add_argument("--high-y")
-    parser.add_argument("--low-z")
-    parser.add_argument("--high-z")
-
-    parser.add_argument("--low-time", default=0)
-    parser.add_argument("--high-time", default=float("inf"))
-
-    # TODO Default
-    #parser.add_argument("-v", "--view")
-    
-    parser.add_argument("-R", "--run", action="store_true")
-
-    args: argparse.Namespace = parser.parse_args()
-
-    # TODO
-    """
-    # Check if the default configuration works/is valid
-    # TODO method to restore default config file for user
-    if os.path.exists(CONFIG_FILE):
-        # parse the config file, check that it's valid
-        # if it's not, either error out, or just ignore it? (Probably error out)
-    """
-    if args.config_file:
-        # This needs to handle everything in the config file, and should be overwritten by any cli switches, so it's fine for this to just go first
-        # Also parse this file. If there is an error in the toml, fail
-        pass
-
-    hdf_source_dir: str
-    if args.hdf_source_directory:
-        hdf_source_dir = os.path.abspath(args.hdf_source_directory)
-
-        if not os.path.exists(hdf_source_dir):
-            print(f"Error: The directory {args.hdf_source_directory} does not exist")
-            sys.exit(1)
-
-    else:
-        hdf_source_dir = os.getcwd()
-
-    odb_source_dir: str
-    if args.odb_source_directory:
-        odb_source_dir = os.path.abspath(args.odb_source_directory)
-
-        if not os.path.exists(odb_source_dir):
-            print(f"Error: The directory {args.odb_source_directory} does not exist")
-            sys.exit(1)
-
-    else:
-        odb_source_dir = os.getcwd()
-
-    results_dir: str = os.path.join(os.getcwd(), "results")
-    if args.results_directory:
-        results_dir = os.path.abspath(args.results_directory)
-
-    if not os.path.exists(results_dir):
-        print(f"Directory {results_dir} does not exist. Creating it now.")
-        os.makedirs(results_dir)
-
-    toml_file: ConfigFiletype = None
-
-    if args.odb:
-        if args.hdf:
-            # Don't allow both an odb and an hdf to be passed in
-            print("Error: You cannot pass in both an .odb file and a .hdf5 file. Please enter one at a time")
-            sys.exit(1)
-
-        else:
-            # Ensure the file exists
-            if not os.path.exists(os.path.join(odb_source_dir, args.odb)):
-                if not os.path.exists(os.path.join(os.getcwd(), args.odb)):
-                    if not os.path.exists(args.odb):
-                        print(f"Error: The file {args.odb} could not be found")
-                        sys.exit(1)
-
-                    else:
-                        state.odb_file = args.odb
-                else:
-                    state.odb_file = os.path.join(os.getcwd(), args.odb)
-            else:
-                state.odb_file = os.path.join(odb_source_dir, args.odb)
-
-        if args.meltpoint:
-            state.meltpoint = args.meltpoint
-
-        if args.low_temp:
-            state.low_temp = args.low_temp
-
-        if args.seed_size:
-            state.mesh_seed_size = args.seed_size
-
-        if args.time_sample:
-            state.time_sample = args.time_sample
-
-        # TODO???
-        #print("Converting this .odb file to a .hdf file")
-        #default: str = os.path.join(hdf_source_dir, state.odb_file.split(os.sep)[-1].split(".")[0] + ".hdf5")
-        #hdf_file_path: str
-        #while True:
-        #    user_input = input(f"Please enter the name for this generated hdf file, or leave blank for the default: {default}")
-        #    if user_input == "":
-        #        user_input = default
-        #    
-        #    if(confirm(f"You entered {user_input}", "Is this correct", "yes")):
-        #        hdf_file_path = os.path.join(hdf_source_dir, user_input)
-        #        break
-
-        #state.odb_to_hdf(hdf_file_path)
-        #toml_file = os.path.join(hdf_source_dir, state.hdf_file.split(".")[0] + ".toml")
-
-    elif args.hdf:
-        # ensure the file exists
-        if not os.path.exists(os.path.join(hdf_source_dir, args.hdf)):
-            if not os.path.exists(os.path.join(os.getcwd(), args.hdf)):
-                if not os.path.exists(args.hdf):
-                    print(f"Error: the file {args.hdf} could not be found")
-                    sys.exit(1)
-
-                else:
-                    state.hdf_file = args.hdf
-            else:
-                state.hdf_file = os.path.join(os.getcwd(), args.hdf)
-        else:
-            state.hdf_file = os.path.join(hdf_source_dir, args.hdf)
-
-        # Search for the stored toml values for this hdf
-        toml_file = os.path.join(hdf_source_dir, state.hdf_file.split(".")[0] + ".toml")
-        if not os.path.exists(toml_file):
-            print(f".toml config file for {state.hdf_file} could not be found")
-
-        if args.meltpoint:
-            state.meltpoint = args.meltpoint
-
-        if args.low_temp:
-            state.low_temp = args.low_temp
-
-        if args.seed_size:
-            state.mesh_seed_size = args.seed_size
-
-        if args.time_sample:
-            state.time_sample = args.time_sample
- 
-    else:
-        # Only if these other switches are used
-        if args.meltpoint or args.low_temp or args.seed_size or args.time_sample:
-            print("INFO: neither a .odb file or a .hdf5 file were provided. You must provide one manually")
-
-            if args.meltpoint:
-                state.meltpoint = args.meltpoint
-
-            if args.low_temp:
-                state.low_temp = args.low_temp
-
-            if args.seed_size:
-                state.mesh_seed_size = args.seed_size
-
-            if args.time_sample:
-                state.time_sample = args.time_sample
-
-    # The 4 dimensions and the two booleans are always set here if they exist
-    if args.low_x:
-        state.set_x_low(args.low_x)
-    if args.high_x:
-        state.set_x_high(args.high_x)
-
-    if args.low_y:
-        state.set_y_low(args.low_y)
-    if args.high_y:
-        state.set_y_high(args.high_y)
-
-    if args.low_z:
-        state.set_z_low(args.low_z)
-    if args.high_z:
-        state.set_z_high(args.high_z)
-
-    if args.low_time:
-        state.set_time_low(args.low_time)
-    if args.high_time:
-        state.set_time_high(args.high_time)
-
-    image_title: str
-    image_label: str
-    if args.title:
-        image_title = args.title
-    else:
-        if hasattr(state, "hdf_file"):
-            image_title = state.hdf_file.split(os.sep)[-1].split(".")[0]
-
-        elif hasattr(state, "odb_file"):
-            image_title = state.odb_file.split(os.sep)[-1].split(".")[0]
-
-        else:
-            image_title = ""
-
-    if args.label:
-        image_label = args.label
-    else:
-        image_label = image_title
-
-    # Manage the final user state and return it
-    user_options: UserOptions = UserOptions(hdf_source_dir, odb_source_dir, results_dir, image_title, image_label, toml_file, args.run)
-
-    return user_options
-
-
 def load_hdf(state: OdbVisualizer):
     state.process_hdf()
-
-def print_state(state: OdbVisualizer, user_options: UserOptions) -> None:
-    print(
-        f"""Current State:
-.hdf5 file:              {state.hdf_file if hasattr(state, "hdf_file") else "not set"}
-.odb file:               {state.odb_file if hasattr(state, "odb_file") else "not set"}
-
-X Range:                 {state.x.low if hasattr(state.x, "low") else "not set"} to {state.x.high - state.mesh_seed_size if hasattr(state.x, "high") and hasattr(state, "mesh_seed_size") else "not set"}
-Y Range:                 {state.y.low if hasattr(state.y, "low") else "not set"} to {state.y.high - state.mesh_seed_size if hasattr(state.y, "high") and hasattr(state, "mesh_seed_size") else "not set"}
-Z Range:                 {state.z.low if hasattr(state.z, "low") else "not set"} to {state.z.high - state.mesh_seed_size if hasattr(state.z, "high") and hasattr(state, "mesh_seed_size") else "not set"}
-Time Range:              {state.time_low if hasattr(state, "time_low") else "not set"} to {state.time_high if hasattr(state, "time_high") else "not set"}
-Seed Size of the Mesh:   {state.mesh_seed_size if hasattr(state, "mesh_seed_size") else "not set"}
-View Angle:              {state.angle if hasattr(state, "angle") else "not set"}
-View Elevation:          {state.elev if hasattr(state, "elev") else "not set"}
-View Azimuth:            {state.azim if hasattr(state, "azim") else "not set"}
-View Roll:               {state.roll if hasattr(state, "roll") else "not set"}
-
-Is each time-step being shown in the PyVista interactive viewer? {"Yes" if state.interactive else "No"}
-Image Title:             {user_options.image_title}
-Image Label:             {user_options.image_label}
-
-Data loaded into memory: {"Yes" if state.loaded else "No"}
-Source directory of .hdf5 files: {user_options.hdf_source_directory}
-Source directory of .odb files: {user_options.odb_source_directory}
-Directory to store results: {user_options.results_directory}"""
-
-    )
 
 
 def set_views(state):
